requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
lxml==4.6.3
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
statement-parser @ file:///Users/jamie/projects/jamie/statement-parser
urllib3==1.26.6
tabulate==0.8.9
py-xbrl==2.0.7
sec_edgar_downloader==4.2.2
pyperclip==1.8.2
secedgar==0.3.3
